Massive patch again
+ bug fixes
+ notification sounds for AI buddy mode.
+ Add hide parameter to the GetSpending command
+ Add general help document
Help doc:
https://micahb.dev/ObsidianViewer.html?file=Markdown/General/General/GPT3%20Bot%20Manual.md
+ TTS help updated
+ JudgeNSFW image support and moderation.
+ Podcast mode updates
+ Volume support for screaming in TTS
    + Add configuration commands for shouting.
+ Move Security_Commands to Config_Commands
+ Add File function to the AI.
+ Add Music function to the AI, using neighboring Catgirl Engineer Gaming music bot I made a long time ago.
+ Add AI controlled RAG. Using Recall and Remember.
    + Ignore Memories.json file used to store info.
+ Add AI controlled delayed functions. Using Timer
+ Add automatic CFG scaling. This enhances AI image quality.
+ Extract AI parameters for easy maintenance.
    + Update to GPT-4o. Now considering renaming AI.
+ Extract pricing to a separate file.
+ Migrate fetchRootBase to User file to fix a circular reference error.
+ Change MusicGen model.
+ Extract VoiceLong to a separate file.
+ Keep VoiceV2 from starting the Python server twice.

+++ VoiceV2.js
@@ -107,37 +107,43 @@ function postJSON(URL, data) {
  */
 let pythonProcess;
 
+let startPromise = null;
+
 /**
  * Starts the AI embedding and voicing server process.
  * @returns {Promise} A promise which resolves when the AI is running.
  */
 function Start() {
-    return new Promise(async res => {
-        if (Started) await Stop();
-
-        // Run the python server.
-        pythonProcess = spawn('python', ['voice_server.py']);
-
-        // Handle process exit events to ensure the Python server is killed when Node.js exits
-        process.on('exit', () => {
-            pythonProcess.kill();
-        });
-
-        pythonProcess.stderr.on('data', (data) => {
-            if (DoConsoleLog)    
-                console.log(data.toString());
-            
-            if (data.includes("Running on")) {
-                Started = true;
-                res();
-            }
-        });
-
-        if (DoConsoleLog)    
-            pythonProcess.stdout.on('data', (data) => {
-                    console.log(data.toString());
+    if (startPromise == null) {
+        startPromise = new Promise(async res => {
+            if (Started) await Stop();
+    
+            // Run the python server.
+            pythonProcess = spawn('python', ['voice_server.py']);
+    
+            // Handle process exit events to ensure the Python server is killed when Node.js exits
+            process.on('exit', () => {
+                pythonProcess.kill();
             });
-    })
+    
+            pythonProcess.stderr.on('data', (data) => {
+                if (DoConsoleLog)    
+                    console.log(data.toString());
+                
+                if (data.includes("Running on")) {
+                    Started = true;
+                    startPromise = null;
+                    res();
+                }
+            });
+    
+            if (DoConsoleLog)    
+                pythonProcess.stdout.on('data', (data) => {
+                        console.log(data.toString());
+                });
+        })
+        return startPromise;
+    } else return startPromise;
 }
 
 function Stop() {
